+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/variables/editor/types/ConstantTypeViewState.kt
@@ -5,4 +5,5 @@ import androidx.compose.runtime.Stable
 @Stable
 data class ConstantTypeViewState(
     val value: String,
+    val isSecret: Boolean,
 ) : VariableTypeViewState