+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/data/domains/import_export/ImportRepository.kt
@@ -480,6 +480,7 @@ constructor(
                 isShareTitle = variable.isShareTitle == true,
                 isMultiline = variable.isMultiline == true,
                 isExcludeValueFromExport = variable.isExcludeValueFromExport == true,
+                isSecret = variable.isSecret == true,
                 sortingOrder = index,
             )
         }